+++ Skin/Retail/FrameXML/DressUpFrames.lua
@@ -6,6 +6,7 @@ if not private.isRetail then return end
 
 --[[ Core ]]
 local Aurora = private.Aurora
+local Base = Aurora.Base
 local Skin = Aurora.Skin
 local Util = Aurora.Util
 
@@ -48,22 +49,51 @@ function private.FrameXML.DressUpFrames()
     -- DressUpFrame --
     ------------------
     local DressUpFrame = _G.DressUpFrame
+
     Skin.ButtonFrameTemplateMinimizable(DressUpFrame)
     Skin.WardrobeOutfitDropDownTemplate(DressUpFrame.OutfitDropDown)
     Skin.MaximizeMinimizeButtonFrameTemplate(DressUpFrame.MaxMinButtonFrame)
 
-    DressUpFrame.ModelScene:SetPoint("TOPLEFT")
-    DressUpFrame.ModelScene:SetPoint("BOTTOMRIGHT")
-
-    DressUpFrame.ModelBackground:SetDrawLayer("BACKGROUND", 3)
-    DressUpFrame.ModelBackground:SetAlpha(0.6)
-
     Skin.UIPanelButtonTemplate(_G.DressUpFrameCancelButton)
-    _G.DressUpFrameCancelButton:SetFrameLevel(_G.DressUpFrameCancelButton:GetFrameLevel() + 1)
+
+    local ModelScene = DressUpFrame.ModelScene
+    ModelScene:SetPoint("TOPLEFT")
+    ModelScene:SetPoint("BOTTOMRIGHT")
+
+    local detailsButton = DressUpFrame.ToggleOutfitDetailsButton
+    Base.CropIcon(detailsButton:GetNormalTexture())
+    Base.CropIcon(detailsButton:GetPushedTexture())
+
+    local settings = private.CLASS_BACKGROUND_SETTINGS[private.charClass.token] or private.CLASS_BACKGROUND_SETTINGS["DEFAULT"];
+    local OutfitDetailsPanel = DressUpFrame.OutfitDetailsPanel
+    local blackBG, classBG, frameBG = OutfitDetailsPanel:GetRegions()
+    blackBG:SetPoint("TOPLEFT", 10, -19)
+    classBG:SetPoint("TOPLEFT", blackBG, 1, -1)
+    classBG:SetPoint("BOTTOMRIGHT", blackBG, -1, 1)
+    classBG:SetDesaturation(settings.desaturation)
+    classBG:SetAlpha(settings.alpha)
+    frameBG:Hide()
+
     Skin.UIPanelButtonTemplate(DressUpFrame.ResetButton)
-    DressUpFrame.ResetButton:SetFrameLevel(DressUpFrame.ResetButton:GetFrameLevel() + 1)
     Util.PositionRelative("BOTTOMRIGHT", DressUpFrame, "BOTTOMRIGHT", -15, 15, 5, "Left", {
         _G.DressUpFrameCancelButton,
         DressUpFrame.ResetButton,
     })
+
+    Skin.UIPanelButtonTemplate(DressUpFrame.LinkButton)
+    DressUpFrame.LinkButton:SetPoint("BOTTOMLEFT", 15, 15)
+
+    DressUpFrame.ModelBackground:SetDrawLayer("BACKGROUND", 3)
+    DressUpFrame.ModelBackground:SetDesaturation(settings.desaturation)
+    DressUpFrame.ModelBackground:SetAlpha(settings.alpha)
+
+
+    -- Raise the frame level of interactable child frames above the model frame.
+    local newFrameLevel = ModelScene:GetFrameLevel() + 1
+    DressUpFrame.OutfitDropDown:SetFrameLevel(newFrameLevel)
+    DressUpFrame.MaximizeMinimizeFrame:SetFrameLevel(newFrameLevel)
+    _G.DressUpFrameCancelButton:SetFrameLevel(newFrameLevel)
+    DressUpFrame.ToggleOutfitDetailsButton:SetFrameLevel(newFrameLevel)
+    DressUpFrame.ResetButton:SetFrameLevel(newFrameLevel)
+    DressUpFrame.LinkButton:SetFrameLevel(newFrameLevel)
 end